Move to one line.
@tjchokel, what do you think of softening the 80 character rule for our
coding style? I've been relaxing it in cases where it's close to 80
chars and line breaking is harding to read.

The main reason to not do this is because if you have to SSH into an 80
character width server to fix something at 4 in the morning you want
your code readable. With Heroku, we don't really have this problem.